+++ device_frame/lib/device_frame.dart
@@ -2,6 +2,11 @@ library device_frame;
 
 export 'src/devices/devices.dart';
 export 'src/frame.dart';
+export 'src/info/identifier.dart';
 export 'src/info/info.dart';
 export 'src/keyboard/virtual_keyboard.dart';
 export 'src/theme.dart';
+export 'src/devices/generic/desktop_monitor/device.dart';
+export 'src/devices/generic/laptop/device.dart';
+export 'src/devices/generic/phone/device.dart';
+export 'src/devices/generic/tablet/device.dart';