+++ device_frame/lib/src/devices/android/samsung_s20/device.dart
@@ -7,26 +7,26 @@ part 'screen.g.dart';
 
 final info = DeviceInfo(
   identifier: const DeviceIdentifier(
-    TargetPlatform.iOS,
+    TargetPlatform.android,
     DeviceType.phone,
-    'iphone-12-mini',
+    'samsung-s20',
   ),
-  name: 'iPhone 12 Mini',
-  pixelRatio: 2.0,
-  frameSize: const Size(871.0, 1768.0),
-  screenSize: const Size(375.0, 812.0),
+  name: 'Samsung Galaxy S20',
+  pixelRatio: 4.0,
   safeAreas: const EdgeInsets.only(
     left: 0.0,
-    top: 44.0,
+    top: 32.0,
     right: 0.0,
-    bottom: 34.0,
+    bottom: 32.0,
   ),
   rotatedSafeAreas: const EdgeInsets.only(
-    left: 44.0,
+    left: 32.0,
     top: 0.0,
-    right: 44.0,
-    bottom: 21.0,
+    right: 32.0,
+    bottom: 0.0,
   ),
   framePainter: const _FramePainter(),
   screenPath: _screenPath,
+  frameSize: const Size(856.54, 1899.0),
+  screenSize: const Size(360.0, 800.0),
 );