+++ device_frame/lib/src/info/info.freezed.dart
@@ -26,7 +26,6 @@ class _$DeviceInfoTearOff {
       required double pixelRatio,
       required CustomPainter framePainter,
       required Size frameSize,
-      required int year,
       required Size screenSize}) {
     return _DeviceInfo(
       identifier: identifier,
@@ -37,7 +36,6 @@ class _$DeviceInfoTearOff {
       pixelRatio: pixelRatio,
       framePainter: framePainter,
       frameSize: frameSize,
-      year: year,
       screenSize: screenSize,
     );
   }
@@ -72,9 +70,6 @@ mixin _$DeviceInfo {
   /// The frame size in pixels.
   Size get frameSize => throw _privateConstructorUsedError;
 
-  /// The year the device has been first released.
-  int get year => throw _privateConstructorUsedError;
-
   /// The size in points of the screen content.
   Size get screenSize => throw _privateConstructorUsedError;
 
@@ -97,7 +92,6 @@ abstract class $DeviceInfoCopyWith<$Res> {
       double pixelRatio,
       CustomPainter framePainter,
       Size frameSize,
-      int year,
       Size screenSize});
 }
 
@@ -119,7 +113,6 @@ class _$DeviceInfoCopyWithImpl<$Res> implements $DeviceInfoCopyWith<$Res> {
     Object? pixelRatio = freezed,
     Object? framePainter = freezed,
     Object? frameSize = freezed,
-    Object? year = freezed,
     Object? screenSize = freezed,
   }) {
     return _then(_value.copyWith(
@@ -155,10 +148,6 @@ class _$DeviceInfoCopyWithImpl<$Res> implements $DeviceInfoCopyWith<$Res> {
           ? _value.frameSize
           : frameSize // ignore: cast_nullable_to_non_nullable
               as Size,
-      year: year == freezed
-          ? _value.year
-          : year // ignore: cast_nullable_to_non_nullable
-              as int,
       screenSize: screenSize == freezed
           ? _value.screenSize
           : screenSize // ignore: cast_nullable_to_non_nullable
@@ -182,7 +171,6 @@ abstract class _$DeviceInfoCopyWith<$Res> implements $DeviceInfoCopyWith<$Res> {
       double pixelRatio,
       CustomPainter framePainter,
       Size frameSize,
-      int year,
       Size screenSize});
 }
 
@@ -206,7 +194,6 @@ class __$DeviceInfoCopyWithImpl<$Res> extends _$DeviceInfoCopyWithImpl<$Res>
     Object? pixelRatio = freezed,
     Object? framePainter = freezed,
     Object? frameSize = freezed,
-    Object? year = freezed,
     Object? screenSize = freezed,
   }) {
     return _then(_DeviceInfo(
@@ -242,10 +229,6 @@ class __$DeviceInfoCopyWithImpl<$Res> extends _$DeviceInfoCopyWithImpl<$Res>
           ? _value.frameSize
           : frameSize // ignore: cast_nullable_to_non_nullable
               as Size,
-      year: year == freezed
-          ? _value.year
-          : year // ignore: cast_nullable_to_non_nullable
-              as int,
       screenSize: screenSize == freezed
           ? _value.screenSize
           : screenSize // ignore: cast_nullable_to_non_nullable
@@ -266,7 +249,6 @@ class _$_DeviceInfo with DiagnosticableTreeMixin implements _DeviceInfo {
       required this.pixelRatio,
       required this.framePainter,
       required this.frameSize,
-      required this.year,
       required this.screenSize});
 
   @override
@@ -304,16 +286,12 @@ class _$_DeviceInfo with DiagnosticableTreeMixin implements _DeviceInfo {
   final Size frameSize;
   @override
 
-  /// The year the device has been first released.
-  final int year;
-  @override
-
   /// The size in points of the screen content.
   final Size screenSize;
 
   @override
   String toString({DiagnosticLevel minLevel = DiagnosticLevel.info}) {
-    return 'DeviceInfo(identifier: $identifier, name: $name, rotatedSafeAreas: $rotatedSafeAreas, safeAreas: $safeAreas, screenPath: $screenPath, pixelRatio: $pixelRatio, framePainter: $framePainter, frameSize: $frameSize, year: $year, screenSize: $screenSize)';
+    return 'DeviceInfo(identifier: $identifier, name: $name, rotatedSafeAreas: $rotatedSafeAreas, safeAreas: $safeAreas, screenPath: $screenPath, pixelRatio: $pixelRatio, framePainter: $framePainter, frameSize: $frameSize, screenSize: $screenSize)';
   }
 
   @override
@@ -329,7 +307,6 @@ class _$_DeviceInfo with DiagnosticableTreeMixin implements _DeviceInfo {
       ..add(DiagnosticsProperty('pixelRatio', pixelRatio))
       ..add(DiagnosticsProperty('framePainter', framePainter))
       ..add(DiagnosticsProperty('frameSize', frameSize))
-      ..add(DiagnosticsProperty('year', year))
       ..add(DiagnosticsProperty('screenSize', screenSize));
   }
 
@@ -353,7 +330,6 @@ class _$_DeviceInfo with DiagnosticableTreeMixin implements _DeviceInfo {
                 other.framePainter == framePainter) &&
             (identical(other.frameSize, frameSize) ||
                 other.frameSize == frameSize) &&
-            (identical(other.year, year) || other.year == year) &&
             (identical(other.screenSize, screenSize) ||
                 other.screenSize == screenSize));
   }
@@ -369,7 +345,6 @@ class _$_DeviceInfo with DiagnosticableTreeMixin implements _DeviceInfo {
       pixelRatio,
       framePainter,
       frameSize,
-      year,
       screenSize);
 
   @JsonKey(ignore: true)
@@ -388,7 +363,6 @@ abstract class _DeviceInfo implements DeviceInfo {
       required double pixelRatio,
       required CustomPainter framePainter,
       required Size frameSize,
-      required int year,
       required Size screenSize}) = _$_DeviceInfo;
 
   @override
@@ -423,10 +397,6 @@ abstract class _DeviceInfo implements DeviceInfo {
 
   /// The frame size in pixels.
   Size get frameSize;
-  @override
-
-  /// The year the device has been first released.
-  int get year;
   @override
 
   /// The size in points of the screen content.